Make private declaration selection helpers internal
